requirements.txt
Django==5.0.9
djangorestframework
djangorestframework-simplejwt
django-cors-headers
channels
bcrypt
daphne
Twisted[http2,tls]
celery
redis
django-celery-beat
psycopg2-binary
dj-database-url
python-decouple
Pillow
requests
google-genai
anthropic
weasyprint
django-storages
boto3
stripe
python-magic
django-filter
whitenoise
sentry-sdk
newrelic
